Add files via upload
 Actionable Insights & Recommendations
In this notebook, we explored the relationship between customer demographics, loan features, and
credit risk. We found that there are a number of factors that can contribute to credit risk, including:
1. Around 80.26% of customers have fully paid their Loan Amount. The defaulters are ~ 20%.
From Personal loan business perspective this ratio is high. These 20% will contribute in NPAs
of LoanTap. To reduce the risk of NPAs, LoanTap should add slightly stringent rules to bring
down this ratio to 5% to 6%.
2. LoanTap should provide loans at slightly higher rate than other Banks. This will offset the
risks of defaulters and maintain the profitability of the business.
3. Overall Statistics of the Model:
• Class-wise precision and recall: The model is better at identifying true positives for class
0 (precision of 0.85) than for class 1 (precision of 0.92). This means that the model is more
likely to correctly classify an example as class 0 if it is actually class 0, but it is more likely
to incorrectly classify an example as class 1 if it is actually class 0.
• F1-score: The F1-scores for both classes are similar (around 0.88), which indicates that the
model is performing well for both classes. However, the F1-score for class 0 is slightly higher
than the F1-score for class 1, which is consistent with the observation that the model is better
at identifying true positives for class 0.
• Accuracy: The overall accuracy of the model is 88%, which is a good score. However, it
is important to remember that accuracy can be misleading, especially when dealing with
imbalanced datasets. In this case, the dataset may be imbalanced, with more examples of
class 0 than class 1. This could lead to the model achieving a high accuracy even if it is not
very good at identifying true positives for class 1. Model created has high values for accuracy,
precision, recall & f1-score. This means, this model is a good classifier. Overall, it has good
prediction capability in identifying right customers (which can be easily converted). However
this model has slightly low capability on correctly identifying defaulters. Overall data has
20% defaulters, model is able to predict 10% of them correctly. Using this model, LoanTap
can easily reduce the ration of defaulters in their portfolio.
These findings suggest that lenders should be cautious when lending to younger borrowers, borrowers with lower incomes, borrowers with high debt-to-income ratios, borrowers with poor credit
histories, borrowers who are taking out large loans, borrowers who are taking out loans with long
terms, and borrowers who are being charged high interest rates.